+++ packages/logs/src/domain/logsCollection/runtimeError/runtimeErrorCollection.ts
@@ -1,8 +1,10 @@
 import type { Context, RawError, ClocksState } from '@datadog/browser-core'
-import { trackRuntimeError, Observable } from '@datadog/browser-core'
+import { ErrorSource, trackRuntimeError, Observable } from '@datadog/browser-core'
+import type { RawRuntimeLogsEvent } from '../../../rawLogsEvent.types'
 import type { LogsConfiguration } from '../../configuration'
-import { reportRawError } from '../../reportRawError'
-import type { Sender } from '../../sender'
+import type { LifeCycle } from '../../lifeCycle'
+import { LifeCycleEventType } from '../../lifeCycle'
+import { StatusType } from '../../logger'
 
 export interface ProvidedError {
   startClocks: ClocksState
@@ -13,14 +15,28 @@ export interface ProvidedError {
 
 export function startRuntimeErrorCollection(
   configuration: LogsConfiguration,
-  sender: Sender,
+  lifeCycle: LifeCycle,
   rawErrorObservable = new Observable<RawError>()
 ) {
   if (configuration.forwardErrorsToLogs) {
     trackRuntimeError(rawErrorObservable)
   }
 
-  const rawErrorSubscription = rawErrorObservable.subscribe((rawError) => reportRawError(rawError, sender))
+  const rawErrorSubscription = rawErrorObservable.subscribe((rawError) => {
+    lifeCycle.notify<RawRuntimeLogsEvent>(LifeCycleEventType.RAW_LOG_COLLECTED, {
+      rawLogsEvent: {
+        message: rawError.message,
+        date: rawError.startClocks.timeStamp,
+        error: {
+          kind: rawError.type,
+          origin: ErrorSource.SOURCE, // Todo: Remove in the next major release
+          stack: rawError.stack,
+        },
+        origin: ErrorSource.SOURCE,
+        status: StatusType.error,
+      },
+    })
+  })
 
   return {
     stop: () => {