+++ packages/rum/src/domain/segmentCollection/segment.ts
@@ -8,7 +8,6 @@ let nextId = 0
 
 export class Segment {
   public isFlushed = false
-  public flushReason?: string
   public readonly metadata: SegmentMetadata
 
   private id = nextId++
@@ -76,13 +75,12 @@ export class Segment {
     this.worker.postMessage({ data: `,${JSON.stringify(record)}`, id: this.id, action: 'write' })
   }
 
-  flush(reason?: string) {
+  flush() {
     this.worker.postMessage({
       data: `],${JSON.stringify(this.metadata).slice(1)}\n`,
       id: this.id,
       action: 'flush',
     })
     this.isFlushed = true
-    this.flushReason = reason
   }
 }