+++ test/e2e/scenario/rum/actions.scenario.ts
@@ -1,3 +1,4 @@
+import { withBrowserLogs } from '../../lib/helpers/browser'
 import { createTest, html, waitForServersIdle } from '../../lib/framework'
 import { flushEvents } from '../../lib/helpers/flushEvents'
 
@@ -38,6 +39,9 @@ describe('action collection', () => {
           name: 'click me',
         },
         type: 'click',
+        frustration: {
+          type: [],
+        },
       })
     })
 
@@ -79,9 +83,79 @@ describe('action collection', () => {
           name: 'click me',
         },
         type: 'click',
+        frustration: {
+          type: [],
+        },
       })
 
       expect(resourceEvents.length).toBe(1)
       expect(resourceEvents[0].action!.id).toBe(actionEvents[0].action.id!)
+    })
+
+  createTest('collect an "error click"')
+    .withRum({ trackInteractions: true, enableExperimentalFeatures: ['frustration-signals'] })
+    .withBody(
+      html`
+        <button>click me</button>
+        <script>
+          const button = document.querySelector('button')
+          button.addEventListener('click', () => {
+            button.setAttribute('data-clicked', 'true')
+            throw new Error('Foo')
+          })
+        </script>
+      `
+    )
+    .run(async ({ serverEvents }) => {
+      const button = await $('button')
+      await button.click()
+      await flushEvents()
+      const actionEvents = serverEvents.rumActions
+
+      expect(actionEvents.length).toBe(1)
+      expect(actionEvents[0].action.frustration!.type).toEqual(['error'])
+      expect(actionEvents[0].action.error!.count).toBe(1)
+      await withBrowserLogs((browserLogs) => {
+        expect(browserLogs.length).toEqual(1)
+      })
+    })
+
+  createTest('collect a "dead click"')
+    .withRum({ trackInteractions: true, enableExperimentalFeatures: ['frustration-signals'] })
+    .withBody(html` <button>click me</button> `)
+    .run(async ({ serverEvents }) => {
+      const button = await $('button')
+      await button.click()
+      await flushEvents()
+      const actionEvents = serverEvents.rumActions
+
+      expect(actionEvents.length).toBe(1)
+      expect(actionEvents[0].action.frustration!.type).toEqual(['dead'])
+    })
+
+  createTest('collect multiple frustrations in one action')
+    .withRum({ trackInteractions: true, enableExperimentalFeatures: ['frustration-signals'] })
+    .withBody(
+      html`
+        <button>click me</button>
+        <script>
+          const button = document.querySelector('button')
+          button.addEventListener('click', () => {
+            throw new Error('Foo')
+          })
+        </script>
+      `
+    )
+    .run(async ({ serverEvents }) => {
+      const button = await $('button')
+      await button.click()
+      await flushEvents()
+      const actionEvents = serverEvents.rumActions
+
+      expect(actionEvents.length).toBe(1)
+      expect(actionEvents[0].action.frustration!.type).toEqual(jasmine.arrayWithExactContents(['error', 'dead']))
+      await withBrowserLogs((browserLogs) => {
+        expect(browserLogs.length).toEqual(1)
+      })
     })
 })